LOCC:  * Removed argh, used CLI11.  * Refactored compiler code  * Added support of other games (only in arch)  * Starting support of Hitman: Contracts  * Changed README  * Added more abstraction layers for easier modification later
