+++ Tools/LOCC/README.md
@@ -1,19 +1,38 @@
 LOCC - LOC Compiler & Decompiler
---------------------------------
+================================
 
- Localization compiler & decompiler for Hitman Blood Money
+Localization compiler & decompiler for Hitman Blood Money
  
 Usage
------
+=====
+
+Decompile:
+----------
 
 ```
-LOCC.exe --from=[Path to localization in JSON] --to=[Path to final LOC file] --test=1
+LOCC.exe --from=M13_main.LOC --to=rel/M13_main.JSON --mode=decompile --pretty-json=on
+```
+
+Compile:
+--------
+
+```
+LOCC.exe --from=rel/M13_main.JSON --to=rel/M13_main.LOC
 ```
 
 Options:
+========
 
- * `--from` - path to source JSON file
- * `--to` - path to final LOC file
- * `--test` - iterate over all keys and try to find it via `ResourceCollection::Lookup`. **It takes a lot of time!**
- * `--decompile-to-json` - decompile file from `--from` as LOC and save to specified path in JSON format
- * `--help` - show help info (all other options will be ignored)
+ * `--f`, `--from` - Path to source file
+ * `--t`, `--to` - Path to destination file
+ * `--m`, `--mode` - Specify tool mode. Allowed values:
+    * `compile` - Use tool as compiler. `--from` must be path to **LOC** file. `--to` must be path to JSON
+    * `decompile` - Use tool as decompiler. ``--from` must be path to **JSON** file. `--to` must be path to LOC
+ * `--g`, `--game` - Specify game name. Allowed values: 
+    * `bloodmoney` - Hitman Blood Money LOC format
+    * `contracts` - Hitman Contracts LOC format
+    * `2sa` - Hitman 2 Silent Assassin
+    * `a47` - Hitman Agent 47
+ * `--p`, `--pretty`, `--pretty-json` - Specify JSON pretty printing. Allowed values:
+    * `on` - Enable JSON pretty printing
+    * `off` - Disable JSON pretty printing (default)
