+++ Tools/LOCC/include/LOCC.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <map>
 #include <string>
 #include <vector>
 #include <fstream>
