+++ Tools/LOCC/include/ResourceCollection.h
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace LOCC
+{
+    struct ResourceCollection
+    {
+    public:
+        static char* Lookup(char* key, char* buffer);
+    };
+}
